Deblurring/requirements.txt
fastapi[all]
natsort==8.1.0
numpy==1.23.1
opencv_python_headless==4.6.0.66
Pillow==9.2.0
requests==2.28.1
scikit_image==0.19.3
thop==0.1.1.post2207130030
torch==1.12.1
torchvision==0.13.1
